fix(slack): Validate channel history timestamps
Validate Slack channel history range bounds before calling conversations.history. Recover Junior Slack thread references only when they match the active destination channel, so malformed or cross-channel values return actionable tool errors instead of Slack API failures.

Fixes #754
