+++ packages/junior/src/chat/slack/tools/list/update-item.ts
@@ -1,0 +1,80 @@
+import { updateListItem } from "@/chat/slack/tools/list/api";
+import { tool } from "@/chat/tools/definition";
+import { createOperationKey } from "@/chat/tools/idempotency";
+import type { ToolState } from "@/chat/tools/types";
+import { Type } from "@sinclair/typebox";
+
+/** Create a tool that updates an item in the active Slack list. */
+export function createSlackListUpdateItemTool(state: ToolState) {
+  return tool({
+    description:
+      "Update an item in the active Slack list tracked in artifact context (title/completion). Use when the user asks to mark progress or rename a tracked task. Do not use to add new tasks.",
+    inputSchema: Type.Object(
+      {
+        item_id: Type.String({
+          minLength: 1,
+          description: "ID of the Slack list item to update.",
+        }),
+        completed: Type.Optional(
+          Type.Boolean({
+            description: "Optional completion status update.",
+          }),
+        ),
+        title: Type.Optional(
+          Type.String({
+            minLength: 1,
+            description: "Optional new item title.",
+          }),
+        ),
+      },
+      {
+        anyOf: [{ required: ["completed"] }, { required: ["title"] }],
+      },
+    ),
+    execute: async ({ item_id, completed, title }) => {
+      const targetListId = state.getCurrentListId();
+      if (!targetListId) {
+        return { ok: false, error: "No active list found in artifact context" };
+      }
+      const operationKey = createOperationKey("slackListUpdateItem", {
+        list_id: targetListId,
+        item_id,
+        completed: completed ?? null,
+        title: title ?? null,
+      });
+      const cached = state.getOperationResult<{
+        ok: true;
+        list_id: string;
+        item_id: string;
+        completed?: boolean;
+        title?: string;
+      }>(operationKey);
+      if (cached) {
+        return {
+          ...cached,
+          deduplicated: true,
+        };
+      }
+
+      await updateListItem({
+        listId: targetListId,
+        itemId: item_id,
+        completed,
+        title,
+        listColumnMap: state.artifactState.listColumnMap ?? {},
+      });
+
+      await state.patchArtifactState({ lastListId: targetListId });
+
+      const response = {
+        ok: true,
+        list_id: targetListId,
+        item_id,
+        completed,
+        title,
+      };
+      state.setOperationResult(operationKey, response);
+      return response;
+    },
+  });
+}
