+++ packages/junior/tests/integration/slack-canvases.test.ts
@@ -1,5 +1,5 @@
 import { beforeEach, describe, expect, it } from "vitest";
-import { createCanvas } from "@/chat/tools/slack/canvases";
+import { createCanvas } from "@/chat/slack/tools/canvas/api";
 import {
   canvasesAccessSetOk,
   canvasesCreateOk,
