+++ packages/junior/tests/integration/slack-list-tools.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from "vitest";
-import { createSlackListGetItemsTool } from "@/chat/tools/slack/list-tools";
+import { createSlackListGetItemsTool } from "@/chat/slack/tools/list/get-items";
 import type { ToolState } from "@/chat/tools/types";
 import { slackListsItemsListPage } from "../fixtures/slack/factories/api";
 import {
