+++ packages/junior/tests/unit/slack/slack-canvas-id.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from "vitest";
-import { extractCanvasId } from "@/chat/tools/slack/canvases";
+import { extractCanvasId } from "@/chat/slack/tools/canvas/api";
 
 describe("extractCanvasId", () => {
   it("returns an uppercased F-prefixed ID as-is", () => {
