+++ packages/junior/tests/unit/slack/slack-message-add-reaction-tool.test.ts
@@ -1,8 +1,8 @@
 import { describe, expect, it, vi } from "vitest";
 import { createSlackSource } from "@sentry/junior-plugin-api";
-import { createSlackMessageAddReactionTool } from "@/chat/tools/slack/message-add-reaction";
+import { createSlackMessageAddReactionTool } from "@/chat/slack/tools/message-add-reaction";
 import { ToolInputError } from "@/chat/tools/execution/tool-input-error";
-import type { SlackToolContext } from "@/chat/tools/slack/context";
+import type { SlackToolContext } from "@/chat/slack/tools/context";
 import { parseSlackMessageTs } from "@/chat/slack/timestamp";
 
 const addReactionToMessage = vi.fn();
